Fix compilation issues after rebase

+++ include/qrw/FootTrajectoryGenerator.hpp
@@ -86,7 +86,7 @@ class FootTrajectoryGenerator {
   double lockTime_;   // Target lock before the touchdown
   double vertTime_;   // Duration during which feet move only along Z when taking off and landing
 
-  Eigen::Matrix<int, 1, 4> feet;  // Column indexes of feet currently in swing phase
+  Eigen::Matrix<double, 1, 4> feet;  // Column indexes of feet currently in swing phase
   Vector4 t0s;                    // Elapsed time since the start of the swing phase movement
   Vector4 t_swing;                // Swing phase duration for each foot
   Vector4 stepHeight_;            // Height of incoming landing point
